+++ packages/netlify-cms-core/src/actions/entries.ts
@@ -20,6 +20,7 @@ import {
   EntryField,
   SortDirection,
   ViewFilter,
+  ViewGroup,
   Entry,
 } from '../types/redux';
 
@@ -31,7 +32,7 @@ import { selectIsFetching, selectEntriesSortFields, selectEntryByPath } from '..
 import { selectCustomPath } from '../reducers/entryDraft';
 import { navigateToEntry } from '../routing/history';
 import { getProcessSegment } from '../lib/formatters';
-import { hasI18n, serializeI18n } from '../lib/i18n';
+import { hasI18n, duplicateDefaultI18nFields, serializeI18n, I18N, I18N_FIELD } from '../lib/i18n';
 
 const { notifSend } = notifActions;
 
@@ -53,6 +54,10 @@ export const SORT_ENTRIES_FAILURE = 'SORT_ENTRIES_FAILURE';
 export const FILTER_ENTRIES_REQUEST = 'FILTER_ENTRIES_REQUEST';
 export const FILTER_ENTRIES_SUCCESS = 'FILTER_ENTRIES_SUCCESS';
 export const FILTER_ENTRIES_FAILURE = 'FILTER_ENTRIES_FAILURE';
+
+export const GROUP_ENTRIES_REQUEST = 'GROUP_ENTRIES_REQUEST';
+export const GROUP_ENTRIES_SUCCESS = 'GROUP_ENTRIES_SUCCESS';
+export const GROUP_ENTRIES_FAILURE = 'GROUP_ENTRIES_FAILURE';
 
 export const DRAFT_CREATE_FROM_ENTRY = 'DRAFT_CREATE_FROM_ENTRY';
 export const DRAFT_CREATE_EMPTY = 'DRAFT_CREATE_EMPTY';
@@ -237,6 +242,44 @@ export function filterByField(collection: Collection, filter: ViewFilter) {
         payload: {
           collection: collection.get('name'),
           filter,
+          error,
+        },
+      });
+    }
+  };
+}
+
+export function groupByField(collection: Collection, group: ViewGroup) {
+  return async (dispatch: ThunkDispatch<State, {}, AnyAction>, getState: () => State) => {
+    const state = getState();
+    const isFetching = selectIsFetching(state.entries, collection.get('name'));
+    dispatch({
+      type: GROUP_ENTRIES_REQUEST,
+      payload: {
+        collection: collection.get('name'),
+        group,
+      },
+    });
+    if (isFetching) {
+      return;
+    }
+
+    try {
+      const entries = await getAllEntries(state, collection);
+      dispatch({
+        type: GROUP_ENTRIES_SUCCESS,
+        payload: {
+          collection: collection.get('name'),
+          group,
+          entries,
+        },
+      });
+    } catch (error) {
+      dispatch({
+        type: GROUP_ENTRIES_FAILURE,
+        payload: {
+          collection: collection.get('name'),
+          group,
           error,
         },
       });
@@ -669,6 +712,9 @@ const processValue = (unsafe: string) => {
   return escapeHtml(unsafe);
 };
 
+const getDataFields = (fields: EntryFields) => fields.filter(f => !f!.get('meta')).toList();
+const getMetaFields = (fields: EntryFields) => fields.filter(f => f!.get('meta') === true).toList();
+
 export function createEmptyDraft(collection: Collection, search: string) {
   return async (dispatch: ThunkDispatch<State, {}, AnyAction>, getState: () => State) => {
     const params = new URLSearchParams(search);
@@ -679,8 +725,12 @@ export function createEmptyDraft(collection: Collection, search: string) {
     });
 
     const fields = collection.get('fields', List());
-    const dataFields = createEmptyDraftData(fields.filter(f => !f!.get('meta')).toList());
-    const metaFields = createEmptyDraftData(fields.filter(f => f!.get('meta') === true).toList());
+
+    const dataFields = getDataFields(fields);
+    const data = createEmptyDraftData(dataFields);
+
+    const metaFields = getMetaFields(fields);
+    const meta = createEmptyDraftData(metaFields);
 
     const state = getState();
     const backend = currentBackend(state.config);
@@ -689,11 +739,14 @@ export function createEmptyDraft(collection: Collection, search: string) {
       await waitForMediaLibraryToLoad(dispatch, getState());
     }
 
+    const i18nFields = createEmptyDraftI18nData(collection, dataFields);
+
     let newEntry = createEntry(collection.get('name'), '', '', {
-      data: dataFields,
+      data,
+      i18n: i18nFields,
       mediaFiles: [],
       // eslint-disable-next-line @typescript-eslint/no-explicit-any
-      meta: metaFields as any,
+      meta: meta as any,
     });
     newEntry = await backend.processEntry(state, collection, newEntry);
     dispatch(emptyDraftCreated(newEntry));
@@ -711,7 +764,11 @@ interface DraftEntryData {
     | (string | DraftEntryData | boolean | List<unknown>)[];
 }
 
-export function createEmptyDraftData(fields: EntryFields, withNameKey = true) {
+export function createEmptyDraftData(
+  fields: EntryFields,
+  withNameKey = true,
+  skipField: (field: EntryField) => boolean = () => false,
+) {
   return fields.reduce(
     (
       reduction: DraftEntryData | string | undefined | boolean | List<unknown>,
@@ -719,6 +776,11 @@ export function createEmptyDraftData(fields: EntryFields, withNameKey = true) {
     ) => {
       const acc = reduction as DraftEntryData;
       const item = value as EntryField;
+
+      if (skipField(item)) {
+        return acc;
+      }
+
       const subfields = item.get('field') || item.get('fields');
       const list = item.get('widget') == 'list';
       const name = item.get('name');
@@ -727,8 +789,8 @@ export function createEmptyDraftData(fields: EntryFields, withNameKey = true) {
 
       if (List.isList(subfields)) {
         const subDefaultValue = list
-          ? [createEmptyDraftData(subfields as EntryFields)]
-          : createEmptyDraftData(subfields as EntryFields);
+          ? [createEmptyDraftData(subfields as EntryFields, withNameKey, skipField)]
+          : createEmptyDraftData(subfields as EntryFields, withNameKey, skipField);
         if (!isEmptyDefaultValue(subDefaultValue)) {
           acc[name] = subDefaultValue;
         } else if (list && List.isList(defaultValue) && (defaultValue as List<unknown>).isEmpty()) {
@@ -740,8 +802,8 @@ export function createEmptyDraftData(fields: EntryFields, withNameKey = true) {
 
       if (Map.isMap(subfields)) {
         const subDefaultValue = list
-          ? [createEmptyDraftData(List([subfields as EntryField]), false)]
-          : createEmptyDraftData(List([subfields as EntryField]));
+          ? [createEmptyDraftData(List([subfields as EntryField]), false, skipField)]
+          : createEmptyDraftData(List([subfields as EntryField]), withNameKey, skipField);
         if (!isEmptyDefaultValue(subDefaultValue)) {
           acc[name] = subDefaultValue;
         } else if (list && List.isList(defaultValue) && (defaultValue as List<unknown>).isEmpty()) {
@@ -762,6 +824,19 @@ export function createEmptyDraftData(fields: EntryFields, withNameKey = true) {
     },
     {} as DraftEntryData,
   );
+}
+
+function createEmptyDraftI18nData(collection: Collection, dataFields: EntryFields) {
+  if (!hasI18n(collection)) {
+    return {};
+  }
+
+  const skipField = (field: EntryField) => {
+    return field.get(I18N) !== I18N_FIELD.DUPLICATE && field.get(I18N) !== I18N_FIELD.TRANSLATE;
+  };
+
+  const i18nData = createEmptyDraftData(dataFields, true, skipField);
+  return duplicateDefaultI18nFields(collection, i18nData);
 }
 
 export function getMediaAssets({ entry }: { entry: EntryMap }) {