+++ packages/netlify-cms-core/src/mediaLibrary.ts
@@ -7,7 +7,7 @@ import { getMediaLibrary } from './lib/registry';
 import store from './redux';
 import { configFailed } from './actions/config';
 import { createMediaLibrary, insertMedia } from './actions/mediaLibrary';
-import { MediaLibraryInstance } from './types/redux';
+import { MediaLibraryInstance, State } from './types/redux';
 
 type MediaLibraryOptions = {};
 
@@ -26,6 +26,8 @@ const initializeMediaLibrary = once(async function initializeMediaLibrary(name, 
     );
     store.dispatch(configFailed(err));
   } else {
+    // eslint-disable-next-line @typescript-eslint/ban-ts-ignore
+    // @ts-ignore
     const handleInsert = (url: string) => store.dispatch(insertMedia(url, undefined));
     const instance = await lib.init({ options, handleInsert });
     store.dispatch(createMediaLibrary(instance));
@@ -33,7 +35,7 @@ const initializeMediaLibrary = once(async function initializeMediaLibrary(name, 
 });
 
 store.subscribe(() => {
-  const state = store.getState();
+  const state = store.getState() as State;
   const mediaLibraryName = state.config.getIn(['media_library', 'name']);
   if (mediaLibraryName && !state.mediaLibrary.get('externalLibrary')) {
     const mediaLibraryConfig = state.config.get('media_library').toJS();