+++ packages/netlify-cms-lib-widgets/src/stringTemplate.ts
@@ -1,11 +1,24 @@
 import moment from 'moment';
 import { Map } from 'immutable';
-import { basename, extname, dirname } from 'path';
+import { basename, dirname, extname } from 'path';
 import { get, trimEnd } from 'lodash';
 
+const filters = [
+  { pattern: /^upper$/, transform: (str: string) => str.toUpperCase() },
+  {
+    pattern: /^lower$/,
+    transform: (str: string) => str.toLowerCase(),
+  },
+  {
+    pattern: /^date\('(.+)'\)$/,
+    transform: (str: string, match: RegExpMatchArray) => moment(str).format(match[1]),
+  },
+];
+
 const FIELD_PREFIX = 'fields.';
-const templateContentPattern = '[^}{]+';
-const templateVariablePattern = `{{(${templateContentPattern})}}`;
+const templateContentPattern = '([^}{|]+)';
+const filterPattern = '( \\| ([^}{]+))?';
+const templateVariablePattern = `{{${templateContentPattern}${filterPattern}}}`;
 
 // prepends a Zero if the date has only 1 digit
 function formatDate(date: number) {
@@ -110,6 +123,21 @@ function getExplicitFieldReplacement(key: string, data: Map<string, unknown>) {
   return value;
 }
 
+function getFilterFunction(filterStr: string) {
+  if (filterStr) {
+    let match: RegExpMatchArray | null = null;
+    const filter = filters.find(filter => {
+      match = filterStr.match(filter.pattern);
+      return !!match;
+    });
+
+    if (filter) {
+      return (str: string) => filter.transform(str, match as RegExpMatchArray);
+    }
+  }
+  return null;
+}
+
 export function compileStringTemplate(
   template: string,
   date: Date | undefined | null,
@@ -125,7 +153,7 @@ export function compileStringTemplate(
 
   const compiledString = template.replace(
     RegExp(templateVariablePattern, 'g'),
-    (_, key: string) => {
+    (_full, key: string, _part, filter: string) => {
       let replacement;
       const explicitFieldReplacement = getExplicitFieldReplacement(key, data);
 
@@ -144,6 +172,11 @@ export function compileStringTemplate(
 
       if (processor) {
         return processor(replacement);
+      } else {
+        const filterFunction = getFilterFunction(filter);
+        if (filterFunction) {
+          replacement = filterFunction(replacement);
+        }
       }
 
       return replacement;