+++ packages/netlify-cms-widget-colorstring/src/index.js
@@ -1,0 +1,12 @@
+import controlComponent from './ColorControl';
+import previewComponent from './ColorPreview';
+
+const Widget = (opts = {}) => ({
+  name: 'color',
+  controlComponent,
+  previewComponent,
+  ...opts,
+});
+
+export const NetlifyCmsWidgetColorString = { Widget, controlComponent, previewComponent };
+export default NetlifyCmsWidgetColorString;