+++ website/src/components/markdownify.js
@@ -2,7 +2,7 @@ import React, { Fragment } from 'react';
 import Markdown from 'react-markdown';
 
 const Markdownify = ({ source }) => (
-  <Markdown source={source} renderers={{ root: Fragment, paragraph: Fragment }} />
+  <Markdown renderers={{ root: Fragment, paragraph: Fragment }}>{source}</Markdown>
 );
 
 export default Markdownify;