+++ packages/netlify-cms/CHANGELOG.md
@@ -3,6 +3,46 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.10.68](https://github.com/netlify/netlify-cms/compare/netlify-cms@2.10.67...netlify-cms@2.10.68) (2020-11-08)
+
+**Note:** Version bump only for package netlify-cms
+
+
+
+
+
+## [2.10.67](https://github.com/netlify/netlify-cms/compare/netlify-cms@2.10.66...netlify-cms@2.10.67) (2020-11-02)
+
+**Note:** Version bump only for package netlify-cms
+
+
+
+
+
+## [2.10.66](https://github.com/netlify/netlify-cms/compare/netlify-cms@2.10.65...netlify-cms@2.10.66) (2020-10-26)
+
+**Note:** Version bump only for package netlify-cms
+
+
+
+
+
+## [2.10.65](https://github.com/netlify/netlify-cms/compare/netlify-cms@2.10.64...netlify-cms@2.10.65) (2020-10-25)
+
+**Note:** Version bump only for package netlify-cms
+
+
+
+
+
+## [2.10.64](https://github.com/netlify/netlify-cms/compare/netlify-cms@2.10.63...netlify-cms@2.10.64) (2020-10-20)
+
+**Note:** Version bump only for package netlify-cms
+
+
+
+
+
 ## [2.10.63](https://github.com/netlify/netlify-cms/compare/netlify-cms@2.10.62...netlify-cms@2.10.63) (2020-10-12)
 
 **Note:** Version bump only for package netlify-cms